frontend/chat/requirements.txt
# Streamlit 聊天服務依賴
streamlit>=1.28.0
aiohttp>=3.9.0
asyncio>=3.4.3
pandas>=2.1.0
plotly>=5.17.0
numpy>=1.24.0
python-dotenv>=1.0.0
edge-tts>=6.1.0
pydantic>=2.5.0
dataclasses-json>=0.6.0
requests>=2.31.0
asyncio-mqtt>=0.16.0
websockets>=12.0
pymongo>=4.6.0
pymilvus>=2.3.0
openai>=1.6.0
anthropic>=0.7.0
google-generativeai>=0.3.0
google-api-python-client>=2.108.0
google-auth>=2.25.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0

# AI 和代理系統
langchain>=0.1.0
langchain-openai>=0.0.2
crewai>=0.11.0
langfuse>=2.7.0
langgraph>=0.4.0

# 向量處理
sentence-transformers>=2.2.0
transformers>=4.30.0

# 工具和監控
prometheus-client>=0.19.0
httpx>=0.25.0

# 移除 protobuf 版本限制，讓 pip 自動解決 